requirements.txt
requests
html2text
PyGithub